requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
WTForms
Werkzeug
google-generativeai 